refactor(celery): move celery_app + celery_worker out of dash/
Relocates the Celery module surface from depictio/dash/ to depictio/api/
so the worker entrypoint stops living inside the Dash package that is
being removed in a later migration step.

  depictio/dash/celery_app.py     → depictio/api/celery_app.py
  depictio/dash/celery_worker.py  → depictio/api/celery_worker.py

Updates all 13 import sites across api/ and dash/, the docstring in
celery_tasks.py, and the run_celery_worker.sh entrypoint.

What this does NOT do:
- celery_worker.py still imports flask_dispatcher to register Dash app
  background callbacks. That coupling stays until Dash itself is
  deleted (Layer 5).
- The lazy `from depictio.dash.modules.*` imports inside individual task
  bodies (figure / multiqc prerender) are untouched — they only fire
  when those tasks run, and the worker image still ships depictio/dash/
  until Layer 5.
- celery_lock.py is dead code (no callers); left in place for Layer 5.

Verified: both modules import cleanly via uv run; ruff/ty introduce no
new diagnostics.
