+++ depictio/api/celery_app.py
